# Conflicts:
#	compat/src/test/java/io/questdb/compat/InfluxDBClientFailureTest.java
#	core/src/main/resources/io/questdb/bin/darwin-aarch64/libquestdb.dylib
#	core/src/main/resources/io/questdb/bin/darwin-x86-64/libquestdb.dylib
#	core/src/main/resources/io/questdb/bin/linux-aarch64/libquestdb.so
#	core/src/main/resources/io/questdb/bin/linux-x86-64/libquestdb.so
#	core/src/main/resources/io/questdb/bin/windows-x86-64/libquestdb.dll

+++ core/src/test/java/io/questdb/test/cairo/TableNameRegistryTest.java
@@ -702,7 +702,7 @@ public class TableNameRegistryTest extends AbstractCairoTest {
 
         FilesFacade ff = new TestFilesFacadeImpl() {
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public long openRW(LPSZ name, int opts) {
                 if (failReloadNameRegistry.get() && Utf8s.endsWithAscii(name, "tables.d.0")) {
                     return -1;
                 }
