+++ core/src/test/java/io/questdb/test/cairo/wal/WalTableSqlTest.java
@@ -1607,7 +1607,7 @@ public class WalTableSqlTest extends AbstractCairoTest {
             int i = 0;
 
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public long openRW(LPSZ name, int opts) {
                 long fd = super.openRW(name, opts);
                 if (Utf8s.containsAscii(name, "2022-02-25") && i++ == 0) {
                     TestUtils.unchecked(() -> execute("drop table " + newTableName));
@@ -1982,7 +1982,7 @@ public class WalTableSqlTest extends AbstractCairoTest {
     public void testSuspendedTablesTriedOnceOnStart() throws Exception {
         FilesFacade ff = new TestFilesFacadeImpl() {
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public long openRW(LPSZ name, int opts) {
                 if (Utf8s.containsAscii(name, "fail.d")) {
                     return -1;
                 }
